Beta

+++ src/base/filesystem.c
@@ -258,7 +258,7 @@ int filesystem_close(int fd)
     return ret;
 }
 
-long filesystem_read(int read_fd2, void * buffer, int buffer_size)
+ssize_t filesystem_read(int read_fd2, void * buffer, size_t buffer_size)
 {
     ssize_t read_num_bytes = -1;
     ssize_t read_remaining_bytes = buffer_size;
@@ -293,7 +293,7 @@ long filesystem_read(int read_fd2, void * buffer, int buffer_size)
     return buffer_size;
 }
 
-long filesystem_write(int write_fd2, void * buffer, int num_bytes_to_write)
+ssize_t filesystem_write(int write_fd2, void * buffer, size_t num_bytes_to_write)
 {
     ssize_t write_num_bytes = -1;
     ssize_t write_remaining_bytes = num_bytes_to_write;
@@ -304,7 +304,8 @@ long filesystem_write(int write_fd2, void * buffer, int num_bytes_to_write)
         debug_warning("[FILE_POSIX]: read_buffer with NULL buffer\n");
     }
 
-    while (write_remaining_bytes > 0) {
+    while (write_remaining_bytes > 0)
+    {
         /* Write into local file (write_fd2)... */
         write_num_bytes = real_posix_write(write_fd2, write_buffer, write_remaining_bytes);
